fixing tag repo

+++ app/Repositories/Tag/TagRepository.php
@@ -6,29 +6,41 @@ use App\Models\Tag;
 
 class TagRepository implements TagRepositoryInterface
 {
-    public function all()
+    public function all($paginate = null)
     {
+        if ($paginate) {
+            return Tag::paginate($paginate);
+        }
         return Tag::all();
     }
 
-    public function find($id): ?Tag
+    public function find($id)
     {
         return Tag::find($id);
     }
 
-    public function create(array $data): Tag
+    public function create(array $data)
     {
         return Tag::create($data);
     }
 
-    public function update(Tag $tag, array $data): Tag
+    public function update($id, array $data)
     {
-        $tag->update($data);
-        return $tag;
+        $tag = Tag::find($id);
+        if ($tag) {
+            $tag->update($data);
+            return $tag;
+        }
+        return null;
     }
 
-    public function delete(Tag $tag): bool
+    public function delete($id)
     {
-        return $tag->delete();
+        $tag = Tag::find($id);
+        if ($tag) {
+            $tag->delete();
+            return true;
+        }
+        return false;
     }
 }
